ajout de l'edition d'evenements

+++ agenda/api/src/main/groovy/ca/uqam/mgl7361/agenda/Evenement.groovy
@@ -2,7 +2,7 @@ package ca.uqam.mgl7361.agenda
 
 class Evenement {
 
-    Long id
+    int id
     String nom
     Date du
     Date au
